@@ -1,9 +1,9 @@
 {
     "require-dev": {
-        "phpunit/phpunit": "9.5.21",
-        "psalm/plugin-phpunit": "0.17.0",
-        "symfony/phpunit-bridge": "6.1.3",
-        "vimeo/psalm": "4.26.0"
+        "phpunit/phpunit": "9.6.17",
+        "psalm/plugin-phpunit": "0.18.4",
+        "symfony/phpunit-bridge": "7.0.4",
+        "vimeo/psalm": "5.23.1"
     },
     "config": {
         "allow-plugins": {
